tradefair_project/requirements.txt
asgiref==3.9.1
boto3==1.35.29
botocore==1.35.99
certifi==2025.10.5
charset-normalizer==3.4.3
coverage==7.10.7
dj-database-url==3.0.1
Django==4.2.16
django-cors-headers==4.9.0
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jmespath==1.0.1
packaging==25.0
paystackapi==2.1.3
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.10.4
setuptools==75.1.0
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
